pyproject.toml
[project]
name = "simple"
version = "0.1.0"
description = "OpenAI Response API agent with A2A compatibility"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.60.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
test = [
    "deepeval>=0.21.0",
    "pytest-asyncio>=0.24.0",
]
redteam = [
    "garak>=0.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
